Create Interace for puzzle to prepare for the second puzzle
